pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "BiteMate"
version = "0.1.0"
description = "AI-powered, Google ADK multi-agent platform for personalized meal planning and cook-mode."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["ai", "meal-planner", "nutrition", "personalization", "google-adk", "agent"]
dependencies = [
    "arxiv>=2.3.1",
    "google-adk>=1.19.0",
    "ipykernel>=7.1.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.1.0",
    "mcp>=1.22.0",
    "pinecone>=8.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "wikipedia>=1.4.0",
]

authors = [
  { name = "Amgoth Sunil", email = "amgothsunil422@gmail.com" }
]

[project.urls]
"Homepage" = "https://github.com/AmgothSunil"
"Repository" = "https://github.com/AmgothSunil/BiteMate"
"Documentation" = "https://github.com/AmgothSunil/BiteMate/blob/main/README.md"

[project.optional-dependencies]
dev = [
  "pytest>=7.2.0",
  "pytest-asyncio>=0.21.0",
  "black>=23.3.0",
  "isort>=5.12.0",
  "mypy>=0.991",
  "pre-commit>=3.3.0",
  "ruff>=0.12.0"
]

cloud = [
  "boto3>=1.26.0",
  "google-cloud-secret-manager>=2.10.0"
]

[project.scripts]
bitemate-cli = "bitemate.api.cli:main"

[tool.setuptools]
packages = { find = { where = ["src"] } }
include-package-data = true
zip-safe = false

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "node-and-date"

